Fix iOS and tvOS implementation
