+++ Sources/Shared/BlueprintLayout.swift
@@ -34,54 +34,29 @@ open class BlueprintLayout : CollectionViewFlowLayout {
     fatalError("init(coder:) has not been implemented")
   }
 
-  func createHeader(_ indexPath: IndexPath, atX x: CGFloat = 0, atY y: CGFloat = 0) -> LayoutAttributes {
-    let layoutAttribute = LayoutAttributes.init(forSupplementaryViewOfKind: .sectionHeader, with: indexPath)
-    layoutAttribute.zIndex = indexPath.section
-    layoutAttribute.size.height = headerReferenceSize.height
-    layoutAttribute.frame.origin.x = x
-    layoutAttribute.frame.origin.y = y
 
-    return layoutAttribute
-  }
-
-  func createFooter(_ indexPath: IndexPath, atX x: CGFloat = 0, atY y: CGFloat = 0) -> LayoutAttributes {
-    let layoutAttribute = LayoutAttributes.init(forSupplementaryViewOfKind: .sectionFooter, with: indexPath)
-    layoutAttribute.zIndex = indexPath.section
-    layoutAttribute.size.height = footerReferenceSize.height
-    layoutAttribute.frame.origin.x = x
-    layoutAttribute.frame.origin.y = y + sectionInset.bottom
-
-    return layoutAttribute
-  }
 
   open override func prepare() {
     self.contentSize = .zero
     self.layoutAttributes = []
   }
 
-  open override func layoutAttributesForSupplementaryView(ofKind elementKind: NSCollectionView.SupplementaryElementKind, at indexPath: IndexPath) -> LayoutAttributes? {
-    let sectionAttributes = layoutAttributes[indexPath.section]
-    var layoutAttributesResult: LayoutAttributes? = nil
+  
 
-    switch elementKind {
-    case .sectionHeader:
-      layoutAttributesResult = sectionAttributes.filter({ $0.representedElementCategory == .supplementaryView }).first
-    case .sectionFooter:
-      layoutAttributesResult = sectionAttributes.filter({ $0.representedElementCategory == .supplementaryView }).last
-    default:
+  override open func layoutAttributesForItem(at indexPath: IndexPath) -> LayoutAttributes? {
+    guard indexPath.section < layoutAttributes.count else {
       return nil
     }
 
-    return layoutAttributesResult
-  }
+    guard indexPath.item < layoutAttributes[indexPath.section].count else {
+      return nil
+    }
 
-  override open func layoutAttributesForItem(at indexPath: IndexPath) -> LayoutAttributes? {
-    return layoutAttributes[indexPath.section][indexPath.item] ?? nil
+    return layoutAttributes[indexPath.section][indexPath.item]
   }
 
   override open func layoutAttributesForElements(in rect: CGRect) -> LayoutAttributesForElements {
-    return layoutAttributes.flatMap { $0 }
-      .filter { $0.frame.intersects(rect) } ?? []
+    return layoutAttributes.flatMap { $0 }.filter { $0.frame.intersects(rect) }
   }
 
   override open func initialLayoutAttributesForAppearingItem(at itemIndexPath: IndexPath) -> LayoutAttributes? {
