+++ Sources/Shared/VerticalBlueprintLayout.swift
@@ -58,7 +58,7 @@ open class VerticalBlueprintLayout: BlueprintLayout {
 
       for item in 0..<numberOfItemsInSection(section) {
         let indexPath = IndexPath(item: item, section: section)
-        let layoutAttribute = LayoutAttributes(forItemWith: indexPath)
+        let layoutAttribute = LayoutAttributes.init(forCellWith: indexPath)
 
         defer { previousItem = layoutAttribute }
