+++ Sources/macOS/Extensions/NSLayoutAttributes.swift
@@ -1,0 +1,7 @@
+import Cocoa
+
+extension NSCollectionViewLayoutAttributes {
+  convenience init(forCellWith indexPath: IndexPath) {
+    self.init(forItemWith: indexPath)
+  }
+}
